SRP: split CorePlugin to EnhancerPlugin and PlayStatusPlugin
* these 2 plugins have totally different responsibilities and priorities
* splitting them allows users to easily override/disable/reorder any of them
