+++ documentation/manual/main.textile
@@ -148,7 +148,7 @@ h2. Class enhancement
 
 A Play plug-in (i.e. a subclass of @play.PlayPlugin@) may include ‘enhancers’ that modify application classes at runtime, to add functionality. This is how some of Play’s magic works.
 
-The built-in @play.CorePlugin@ uses enhancers from the @play.classloading.enhancers@ package to dynamically add code to your application’s classes:
+The built-in @play.plugins.EnhancerPlugin@ uses enhancers from the @play.classloading.enhancers@ package to dynamically add code to your application’s classes:
 
 * @ContinuationEnhancer@ - adds continuations support to controller classes
 * @ControllersEnhancer@ - makes controller action methods thread-safe and adds HTTP redirects for method calls
