+++ framework/src/play.plugins
@@ -1,4 +1,4 @@
-0:play.CorePlugin
+0:play.plugins.EnhancerPlugin
 1:play.ConfigurationChangeWatcherPlugin
 100:play.data.parsing.TempFilePlugin
 200:play.data.validation.ValidationPlugin
@@ -9,3 +9,4 @@
 600:play.libs.WS
 700:play.jobs.JobsPlugin
 100000:play.plugins.ConfigurablePluginDisablingPlugin
+100100:play.plugins.PlayStatusPlugin
