+++ framework/test-src/play/plugins/custom-play-with-blank-lines.plugins
@@ -1,4 +1,4 @@
-0:play.CorePlugin
+0:play.plugins.EnhancerPlugin
 
 1:play.plugins.TestPlugin
