+++ framework/test-src/play/plugins/custom-play.plugins
@@ -1,2 +1,2 @@
-0:play.CorePlugin
+0:play.plugins.EnhancerPlugin
 1:play.plugins.TestPlugin
